+++ Termino.h
@@ -46,7 +46,7 @@ int DevuelveINT (array_dinamico a, int cant);
 //Potencia enteros
 int potencia(int x, int y);
 //Potencia enteros largos
-long int potenciaLong(long int x, long int y);
+long int potenciaLong(long int x, int y);
 
 //SUMA Percondicion que los exponentes sean iguales
 Termino SumarTermino(Termino T, Termino T2);
